+++ backend/src/models/mod.rs
@@ -6,4 +6,5 @@ pub mod prelude {
     pub use crate::prelude::*;
     pub use hyper::Uri;
     pub use std::{net::SocketAddr, time::Duration};
+    pub use tsync::tsync;
 }