+++ frontend/astro.config.mjs
@@ -1,10 +1,19 @@
 import { defineConfig } from 'astro/config';
 import react from '@astrojs/react';
-
 import tailwind from '@astrojs/tailwind';
+// import node from '@astrojs/node';
 
 // https://astro.build/config
 export default defineConfig({
-  integrations: [react(), tailwind()],
-  outDir: './frontend',
+    // output: 'server',
+    integrations: [
+        react({
+            experimentalReactChildren: true,
+        }),
+        tailwind(),
+    ],
+    // adapter: node({
+    //     mode: 'standalone',
+    // }),
+    outDir: './frontend',
 });