+++ frontend/src/components/Context.ts
@@ -1,0 +1,7 @@
+import { persistentMap } from '@nanostores/persistent';
+
+const context: Context = {
+    refreshTime: '1',
+    enabled: false,
+};
+export const Context = persistentMap<Context>('context', context);