+++ frontend/src/env.d.ts
@@ -1,1 +1,8 @@
 /// <reference types="astro/client" />
+/** Global App Context */
+type Context = {
+    /** Refresh Time in minutes */
+    refreshTime: RefreshTime;
+    /** Is server must fetch data */
+    enabled: bool;
+};